[MOOSE-201]: Move JS into the Blocks folder
