+++ wp-content/themes/core/assets/js/theme/blocks/details.js
